# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Popcorn API",
    "description": "Popcorn API provides a simple and organized backend to store movies, users, genres, and reviews. Fully documented and built with Node.js, Express, and MongoDB."
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "oAuth2": {
      "type": "oauth2",
      "flow": "implicit",
      "authorizationUrl": "http://localhost:3000/auth/google",
      "scopes": {
        "profile": "Access to user profile",
        "email": "Access to user email"
      }
    },
    "apiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "X-API-KEY",
      "description": "Some description..."
    }
  },
  "paths": {
...
    "/auth/login/success": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/auth/logout": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/auth/user": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  }
}